requirements.txt
flask
flask_sqlalchemy
python-dotenv
psycopg2-binary
gunicorn
